% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestbeforemaxdd.R
\name{bestbeforemaxdd}
\alias{bestbeforemaxdd}
\title{Function to calculate the difference before meet the first max drawdown}
\usage{
bestbeforemaxdd(x, threshold, long = TRUE)
}
\arguments{
\item{x}{data series, should be a vector}

\item{threshold}{the threshold to determine max drawdown}

\item{long}{is long or short?}
}
\description{
Function to calculate the difference before meet the first max drawdown
}
\examples{
bestbeforemaxdd()
}
\keyword{maxdd}

